+++ Editor/Playa/Renderer/ButtonFakeRenderer/ButtonRendererUIToolkit.cs
@@ -21,8 +21,8 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
 {
     public partial class ButtonRenderer
     {
-        private static string ButtonName(MethodInfo methodInfo, object target) => $"{target?.GetHashCode()}_{methodInfo.Name}_{string.Join("_", methodInfo.GetParameters().Select(each => each.Name))}__ButtonRenderer";
-        private static string ButtonLabelContainerName(MethodInfo methodInfo, object target) => $"{target?.GetHashCode()}_{methodInfo.Name}_{string.Join("_", methodInfo.GetParameters().Select(each => each.Name))}__ButtonLabelContainer";
+        private static string ButtonName(SaintsFieldWithInfo info) => $"{info.Targets[0]?.GetHashCode()}_{info.MethodInfo.Name}_{string.Join("_", info.MethodInfo.GetParameters().Select(each => each.Name))}__ButtonRenderer";
+        // private static string ButtonLabelContainerName(MethodInfo methodInfo, object target) => $"{target?.GetHashCode()}_{methodInfo.Name}_{string.Join("_", methodInfo.GetParameters().Select(each => each.Name))}__ButtonLabelContainer";
         // private static string ButtonRotatorName(MethodInfo methodInfo, object target) => $"{target?.GetHashCode()}_{methodInfo.Name}_{string.Join("_", methodInfo.GetParameters().Select(each => each.Name))}__ButtonLabelContainer";
 
         private static StyleSheet _ussClassSaintsFieldEditingDisabledHide;
@@ -37,25 +37,40 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
             public List<Waiter> Enumerators = new List<Waiter>();
         }
 
-        private VisualElement _returnValueContainer;
-        private VisualElement _returnContainer;
+        // private VisualElement _returnValueContainer;
+        // private VisualElement _returnContainer;
 
-        private Button _buttonElement;
-        private IVisualElementScheduledItem buttonTask = null;
+        // private Button _buttonElement;
+        private IVisualElementScheduledItem _buttonTask;
 
         protected override (VisualElement target, bool needUpdate) CreateTargetUIToolkit(VisualElement inspectorRoot,
             VisualElement container)
         {
+            FancyButton fancyButton = new FancyButton
+            {
+                name = ButtonName(FieldWithInfo),
+            };
+            // fancyButton.MainLabel.Add(new Label(FieldWithInfo.MethodInfo.Name));
+            // fancyButton.StatusIndicator.EnsureLoading(true, 0.7f);
+            //
+            // fancyButton.MainButton.clicked += () =>
+            // {
+            //     fancyButton.ShowCloseButton(fancyButton.CloseButton.style.display != DisplayStyle.Flex);
+            // };
+
+            // return (fancyButton, true);
             container.style.flexGrow = 1;
+
             MethodInfo methodInfo = FieldWithInfo.MethodInfo;
             // Debug.Assert(methodInfo.GetParameters().All(p => p.IsOptional));
-            string buttonText = string.IsNullOrEmpty(_buttonAttribute.Label) || _buttonAttribute.IsCallback ? ObjectNames.NicifyVariableName(methodInfo.Name) : _buttonAttribute.Label;
+            string buttonText = string.IsNullOrEmpty(_buttonAttribute.Label)
+                                || _buttonAttribute.IsCallback ? ObjectNames.NicifyVariableName(methodInfo.Name) : _buttonAttribute.Label;
             // object[] defaultParams = methodInfo.GetParameters().Select(p => p.DefaultValue).ToArray();
             ParameterInfo[] parameters = methodInfo.GetParameters();
             bool hasParameters = parameters.Length > 0;
             // List<VisualElement> parameterElements = new List<VisualElement>();
             object[] parameterValues = new object[parameters.Length];
-            VisualElement root = null;
+            // VisualElement root = null;
 
             bool hasReturnValue = !_buttonAttribute.HideReturnValue
                 && methodInfo.ReturnType != typeof(void)
@@ -63,36 +78,9 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
 
             if (hasParameters || hasReturnValue)
             {
-                root = new VisualElement
-                {
-                    style =
-                    {
-                        backgroundColor = new Color(64f/255, 64f/255, 64f/255, 1f),
-                        borderTopWidth = 1,
-                        borderLeftWidth = 1,
-                        borderRightWidth = 1,
-                        borderBottomWidth = 1,
-                        borderLeftColor = EColor.MidnightAsh.GetColor(),
-                        borderRightColor = EColor.MidnightAsh.GetColor(),
-                        borderTopColor = EColor.MidnightAsh.GetColor(),
-                        borderBottomColor = EColor.MidnightAsh.GetColor(),
-                        borderTopLeftRadius = 3,
-                        borderTopRightRadius = 3,
-                        borderBottomLeftRadius = 3,
-                        borderBottomRightRadius = 3,
-                        marginTop = 1,
-                        marginBottom = 1,
-                        marginLeft = 3,
-                        marginRight = 3,
-                        paddingTop = 3,
-                    },
-                    name = "saintsfield-button-params",
-                };
+                VisualElement parametersContainer = fancyButton.HasParameters();
 
-                _ussClassSaintsFieldEditingDisabledHide ??= Util.LoadResource<StyleSheet>("UIToolkit/ClassSaintsFieldEditingDisabledHide.uss");
-                root.styleSheets.Add(_ussClassSaintsFieldEditingDisabledHide);
-
-                root.RegisterCallback<AttachToPanelEvent>(_ => UIToolkitUtils.LoopCheckOutOfScoopFoldout(root));
+                parametersContainer.RegisterCallback<AttachToPanelEvent>(_ => UIToolkitUtils.LoopCheckOutOfScoopFoldout(parametersContainer));
 
                 foreach ((ParameterInfo parameterInfo, int index) in parameters.WithIndex())
                 {
@@ -103,7 +91,7 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                             marginRight = 4,
                         },
                     };
-                    root.Add(paraContainer);
+                    parametersContainer.Add(paraContainer);
 
                     Type paraType = parameterInfo.ParameterType;
                     object paraValue;
@@ -136,10 +124,7 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                             {
                                 paraValue = parameterValues[index] = newValue;
                                 paraValueChanged = true;
-                                if(_returnContainer != null && _returnContainer.style.display == DisplayStyle.Flex)
-                                {
-                                    _returnContainer.style.display = DisplayStyle.None;
-                                }
+                                fancyButton.ShowResult(false);
                             },
                             false,
                             InAnyHorizontalLayout,
@@ -166,28 +151,29 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                 UpdateOneMoreTime = true,
                 Enumerators = new List<Waiter>(),
             };
+            fancyButton.userData = buttonUserData;
 
-            // IVisualElementScheduledItem buttonTask = null;
-            StatusIndicatorElement statusIndicatorElement = new StatusIndicatorElement
+            fancyButton.CloseButton.clicked += () =>
             {
-                style =
+                _buttonTask?.Pause();
+
+                fancyButton.StatusIndicator.EnsureLoading(false, 0);
+                if (buttonUserData.Enumerators.Count > 0)
                 {
-                    position = Position.Absolute,
-                    // width = 16,
-                    // height = 16,
-                    // left = 1,
-                    top = 1,
-                    left = 0,
-                    right = 0,
-                    bottom = 0,
-                    // display = DisplayStyle.None,
-                },
+                    fancyButton.StatusIndicator.PlayPause();
+                }
+                buttonUserData.Enumerators.Clear();
+                fancyButton.ShowResult(false);
             };
+
+            StatusIndicatorElement statusIndicatorElement = fancyButton.StatusIndicator;
 
             bool isStruct = ReflectUtils.TypeIsStruct(FieldWithInfo.Targets[0].GetType());
 
-            _buttonElement = new Button(() =>
+            fancyButton.MainButton.clicked += () =>
             {
+                fancyButton.ShowResult(false);
+
                 SaintsContext.SerializedProperty = _serializedProperty;
                 int targetCount = FieldWithInfo.Targets.Count;
                 object[] returnValues = new object[targetCount];
@@ -219,16 +205,20 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                 if (error != null)
                 {
                     statusIndicatorElement.PlayError();
+                    VisualElement resultContainer = fancyButton.ShowResult(true);
+                    resultContainer.Clear();
+                    resultContainer.Add(MakeErrorBox(error));
                     return;
                 }
 
                 if (hasReturnValue)
                 {
-                    Debug.Assert(_returnValueContainer != null);
-                    _returnValueContainer.Clear();
+                    // Debug.Assert(_returnValueContainer != null);
+                    VisualElement returnValueContainer = fancyButton.ShowResult(true);
+                    returnValueContainer.Clear();
                     object returnValue = returnValues[0];
                     VisualElement r = UIToolkitEdit.UIToolkitValueEdit(
-                        _returnValueContainer.Children().FirstOrDefault(),
+                        returnValueContainer.Children().FirstOrDefault(),
                         "<color=green>[return]</color>",
                         methodInfo.ReturnType,
                         returnValue,
@@ -247,11 +237,14 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                             fo.RegisterCallback<AttachToPanelEvent>(_ => fo.value = true);
                             // fo.value = true;
                         }
-                        if(_returnContainer.style.display != DisplayStyle.Flex)
-                        {
-                            _returnContainer.style.display = DisplayStyle.Flex;
-                        }
-                        _returnValueContainer.Add(r);
+
+                        fancyButton.ShowResult(true);
+                        // if (_returnContainer.style.display != DisplayStyle.Flex)
+                        // {
+                        //     _returnContainer.style.display = DisplayStyle.Flex;
+                        // }
+
+                        returnValueContainer.Add(r);
                     }
                 }
 
@@ -269,13 +262,15 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                 else
                 {
                     statusIndicatorElement.PlayLoading();
+                    fancyButton.ShowCloseButton(true);
                 }
-                buttonTask?.Pause();
+
+                _buttonTask?.Pause();
 
                 if (buttonUserData.Enumerators.Count > 0)
                 {
                     // ButtonUserData buttonUserData = (ButtonUserData) buttonElement.userData;
-                    buttonTask = _buttonElement.schedule.Execute(() =>
+                    _buttonTask = fancyButton.schedule.Execute(() =>
                     {
                         List<Waiter> finishedEnumerators = new List<Waiter>();
                         int oldCounter = buttonUserData.Enumerators.Count;
@@ -293,6 +288,7 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                                     float curProcess = waiter.Waitable.Progress;
                                     progress = Mathf.Max(progress, curProcess);
                                 }
+
                                 continue;
                             }
 
@@ -308,6 +304,9 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                                 movingError = e;
                                 moveNext = false;
                                 thisHasMoveError = true;
+
+                                VisualElement result = fancyButton.ShowResult(true);
+                                result.Add(MakeErrorBox(e));
                             }
 
                             if (thisHasMoveError)
@@ -326,97 +325,29 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                             }
                         }
 
-                        if(movingError == null)
-                        {
-                            buttonUserData.Enumerators.RemoveAll(each => finishedEnumerators.Contains(each));
-                        }
-                        else
-                        {
-                            buttonUserData.Enumerators.Clear();
-                        }
+                        buttonUserData.Enumerators.RemoveAll(each => finishedEnumerators.Contains(each));
 
                         bool stillHaveRunner = buttonUserData.Enumerators.Count > 0;
                         statusIndicatorElement.EnsureLoading(stillHaveRunner, progress);
 
-                        if(!stillHaveRunner)
+                        if (!stillHaveRunner)
                         {
-                            // ReSharper disable once AccessToModifiedClosure
-                            // ReSharper disable once PossibleNullReferenceException
-                            buttonTask?.Pause();
+                            _buttonTask?.Pause();
 
                             if (oldCounter > 0)
                             {
-                                if(movingError == null)
-                                {
-                                    statusIndicatorElement.PlayOk();
-                                }
-                                else
-                                {
-                                    statusIndicatorElement.PlayError();
-                                }
+                                statusIndicatorElement.PlayOk();
                             }
                         }
                     }).Every(1);
                 }
-
-                // ReSharper disable once InvertIf
-
-            })
-            {
-                text = "",
-                enableRichText = true,
-                style =
-                {
-                    flexGrow = 1,
-                    flexDirection = FlexDirection.Row,
-                    justifyContent = Justify.Center,
-                    position = Position.Relative,
-                },
-                name = ButtonName(FieldWithInfo.MethodInfo, FieldWithInfo.Targets[0]),
-                userData = buttonUserData,
             };
 
-            // if (!string.IsNullOrEmpty(buttonAttribute.Label))
-            // {
-            //     buttonElement.text = "";
-            //     buttonElement.Clear();
-            //     foreach (VisualElement element in (new RichTextDrawer()).DrawChunksUIToolKit(RichTextDrawer.ParseRichXml(buttonText,
-            //                  FieldWithInfo.MethodInfo.Name, FieldWithInfo.MethodInfo, FieldWithInfo.Target)))
-            //     {
-            //         buttonElement.Add(element);
-            //     }
-            // }
-
-            _buttonElement.Clear();
-            VisualElement buttonLabelContainer = new VisualElement
-            {
-                style =
-                {
-                    flexGrow = 1,
-                    flexDirection = FlexDirection.Row,
-                    justifyContent = Justify.Center,
-                    alignItems = Align.Center,
-                },
-                name = ButtonLabelContainerName(FieldWithInfo.MethodInfo, FieldWithInfo.Targets[0]),
-            };
-            _buttonElement.Add(buttonLabelContainer);
+            fancyButton.MainLabel.Clear();
             foreach (VisualElement element in new RichTextDrawer().DrawChunksUIToolKit(RichTextDrawer.ParseRichXmlWithProvider(buttonText, this)))
             {
-                buttonLabelContainer.Add(element);
+                fancyButton.MainLabel.Add(element);
             }
-
-            // buttonLabelContainer.RegisterCallback<AttachToPanelEvent>(_ =>
-            //     UIToolkitUtils.TriggerRotate(buttonLabelContainer));
-            // UIToolkitUtils.TriggerRotate(buttonLabelContainer);
-            // buttonRotator.transform.rotation = Quaternion.Euler(0, 0, 180);
-            // buttonRotator.AddToClassList("saints-rotate-360");
-
-            // _buttonElement.Add(buttonRotator);
-            _buttonElement.Add(statusIndicatorElement);
-            // _buttonElement.clicked += () =>
-            // {
-            //     statusIndicatorElement.DoPlay();
-            // };
 
             bool needUpdate = _buttonAttribute.IsCallback;
 
@@ -433,50 +364,9 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
             string methodNameFriendly = ObjectNames.NicifyVariableName(methodInfo.Name);
 
             _onSearchFieldUIToolkit.AddListener(Search);
-            _buttonElement.RegisterCallback<DetachFromPanelEvent>(_ => _onSearchFieldUIToolkit.RemoveListener(Search));
+            fancyButton.RegisterCallback<DetachFromPanelEvent>(_ => _onSearchFieldUIToolkit.RemoveListener(Search));
 
-            if (!hasParameters && !hasReturnValue)
-            {
-                return (_buttonElement, needUpdate);
-            }
-            _buttonElement.style.marginTop = _buttonElement.style.marginBottom = _buttonElement.style.marginLeft = _buttonElement.style.marginRight = 0;
-            _buttonElement.style.borderTopLeftRadius = _buttonElement.style.borderTopRightRadius = 0;
-            _buttonElement.style.borderLeftWidth = _buttonElement.style.borderRightWidth = _buttonElement.style.borderBottomWidth = 0;
-            root.Add(_buttonElement);
-
-            // ReSharper disable once InvertIf
-            if(hasReturnValue)
-            {
-                _returnContainer = new VisualElement
-                {
-                    style =
-                    {
-                        display = DisplayStyle.None,
-                        marginRight = 4,
-                    },
-                };
-                _returnValueContainer = new VisualElement();
-                _returnContainer.Add(_returnValueContainer);
-
-                _returnContainer.Add(new Button(() =>
-                {
-                    _returnContainer.style.display = DisplayStyle.None;
-                })
-                {
-                    text = "x",
-                    style =
-                    {
-                        position = Position.Absolute,
-                        top = 0,
-                        right = -4,
-                    },
-                });
-
-                // returnValueContainer.SetEnabled(false);
-                root.Add(_returnContainer);
-            }
-
-            return (root, needUpdate);
+            return (fancyButton, needUpdate);
 
             void Search(string search)
             {
@@ -484,11 +374,32 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                     ? DisplayStyle.Flex
                     : DisplayStyle.None;
 
-                if (_buttonElement.style.display != display)
+                if (fancyButton.style.display != display)
                 {
-                    _buttonElement.style.display = display;
+                    fancyButton.style.display = display;
                 }
             }
+        }
+
+        private static VisualElement MakeErrorBox(Exception error)
+        {
+            return new HelpBox(error.InnerException?.Message ?? error.Message, HelpBoxMessageType.Error)
+            {
+                style =
+                {
+                    borderLeftWidth = 0,
+                    borderRightWidth = 0,
+                    // borderTopWidth = 0,
+                    borderTopLeftRadius = 0,
+                    borderTopRightRadius = 0,
+                    borderBottomWidth = 0,
+                    backgroundColor = Color.clear,
+                    marginTop = 0,
+                    marginBottom = 0,
+                    marginLeft = 0,
+                    marginRight = 0,
+                },
+            };
         }
 
         // private RichTextDrawer _richTextDrawer;
@@ -507,10 +418,10 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                 CheckMethodBind(playaMethodBindAttribute, FieldWithInfo);
             }
 
-            Button buttonElement;
+            FancyButton fancyButton;
             try
             {
-                buttonElement = root.Q<Button>(name: ButtonName(FieldWithInfo.MethodInfo, FieldWithInfo.Targets[0]));
+                fancyButton = root.Q<FancyButton>(name: ButtonName(FieldWithInfo));
             }
             catch (NullReferenceException)
             {
@@ -521,12 +432,12 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
                 return baseResult;
             }
 
-            if (buttonElement == null)
+            if (fancyButton == null)
             {
                 return baseResult;
             }
 
-            ButtonUserData buttonUserData = (ButtonUserData) buttonElement.userData;
+            ButtonUserData buttonUserData = (ButtonUserData) fancyButton.userData;
 
             string labelCallback = buttonUserData.Callback;
             // ReSharper disable once InvertIf
@@ -570,7 +481,7 @@ namespace SaintsField.Editor.Playa.Renderer.ButtonFakeRenderer
 
                 // buttonElement.text = "";
                 // buttonElement.Clear();
-                VisualElement buttonLabelContainer = root.Q<VisualElement>(name: ButtonLabelContainerName(FieldWithInfo.MethodInfo, FieldWithInfo.Targets[0]));
+                VisualElement buttonLabelContainer = fancyButton.MainLabel;
                 buttonLabelContainer.Clear();
 
                 // ReSharper disable once ConvertIfStatementToSwitchStatement